# backend/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: yourfriends@flexiwan.com
    name: API support
    url: https://flexiwan.com
  description: This is the REST API for flexiWAN management.
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://flexiwan.com/legal/terms-of-service/
  title: FlexiWAN API
  version: 1.0.0
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/frozen-eye/flexiwanManage/1.0.0
- description: Production server
  url: https://manage.flexiwan.com/api/v1
- description: Local Development server
  url: https://local.flexiwan.com:3443/api
paths:
...
  /devices:
    get:
      operationId: devicesGET
      parameters:
      - description: The number of items to skip before starting to collect the result
          set
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
      - description: The numbers of items to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Device'
                type: array
                x-content-type: application/json
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Get all registered devices
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
  /devices/latestVersions:
    get:
      operationId: devicesLatestVersionsGET
      responses:
       "200":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/DeviceLatestVersion'
         description: Successful operation
       "401":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unauthorized
       "403":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Forbidden
       "404":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: The specified resource was not found
       "500":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Internal server error
       default:
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Get all registered devices
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
  /devices/upgdSched:
    post:
      operationId: devicesUpgdSchedPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicesUpgradeRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
          description: Created
          headers:
            Location:
              description: Location of a new resource created
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Create new access token
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
  /devices/{id}:
    get:
      operationId: devicesIdGET
      parameters:
      - description: Numeric ID of the Device to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Device'
                type: array
                x-content-type: application/json
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Get all registered devices
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
    delete:
      operationId: devicesIdDELETE
      parameters:
      - description: Numeric ID of the Device to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
       "204":
         description: Successfull operation
       "401":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unauthorized
       "403":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Forbidden
       "404":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: The specified resource was not found
       "500":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Internal server error
       default:
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Delete device
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
    put:
      operationId: devicesIdPUT
      parameters:
      - description: Numeric ID of the Device to modify
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
       content:
         application/json:
           schema:
             $ref: '#/components/schemas/DeviceRequest'
      responses:
       "200":
         content:
           application/json:
             schema:
               items:
                 $ref: '#/components/schemas/Device'
               type: array
               x-content-type: application/json
         description: Successful operation
       "401":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unauthorized
       "403":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Forbidden
       "404":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: The specified resource was not found
       "500":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Internal server error
       default:
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Modify device
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
...
components:
...
  schemas:
...
    DeviceVersion:
      example:
        agent: agent
        router: router
        device: device
      properties:
        agent:
          type: string
        device:
          type: string
        router:
          type: string
      type: object
    DeviceLatestVersion:
      example:
        "versions":
          "device": "1.1.51"
          "agent": "1.0.39"
          "vpp": "19.01-rc0"
          "frr": "7.1"
          "router": "01.01.10"
        "versionDeadline": "2020-03-15T00:00:00.000Z"
      properties:
        versions:
          $ref: '#/components/schemas/DeviceVersion'
        versionDeadLite:
          type: string
...
    DevicesUpgradeRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        date:
          type: string
...
    Error:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
  securitySchemes:
    apiKey:
      in: header
      name: APIKEY
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http